fix: remove unlinked bookmark-compat redirects instead of documenting retention (#843)
Prior commit kept /account, /achievements, /opportunities with a retention-date
comment and a wiki decision-note. Owner feedback: remove the routes outright
and drop the wiki write-up, not defer them. Ported the edit-form coverage that
only reached ProfileOverviewPage via the /account redirect into
ProfileOverviewTests.cs directly against /profile, and pointed the
achievements-tab deep-link test at /profile?tab=achievements instead of the
removed /achievements redirect.
