+++ wiki/bundle/index.md
@@ -16,5 +16,5 @@ cover. See `../AGENTS.md` for when to split a section further.
 - [Process](process/index.md) (5) - Workflows and procedures - releasing, the mandatory deploy-and-verify flow, live Playwright scripts, EF migrations, and keeping this wiki self-building.
 - [Gotchas](gotchas/index.md) (5) - Traps and non-obvious constraints, each learned from fixing a real bug or hitting a real wall.
 - [Reference](reference/index.md) (4) - Stable reference material - the conventions the architecture tests enforce, the frontend stack and its lint gaps, the Keycloak realm, and a pointer to the formal ADRs/TDRs.
-- [Decisions](decisions/index.md) (4) - Why the repo's tooling and autonomous routines are set up the way they are.
+- [Decisions](decisions/index.md) (3) - Why the repo's tooling and autonomous routines are set up the way they are.
 - [CI](ci/index.md) (1) - Recurring CI failure modes and their causes.
