+++ wiki/bundle/log.md
@@ -12,7 +12,6 @@ dates and a bold action-word prefix, e.g. `**Added**`, `**Updated**`,
 
 ## 2026-07-24
 
-- **Added** - `decisions/bookmark-compat-redirect-retention.md`: #843 found three redirect-only routes in `App.tsx` (`/account`, `/achievements`, `/opportunities`) with no in-app link pointing at them. Decision: a 6-month retention window per route, anchored to the restructure that orphaned it, recorded as a remove-after date both in the page and as a code comment above each `<Route>`. `/my-engagements` was checked and excluded - `Header.tsx:104` still links to it.
 - **Fixed** - `reference/adr-tdr-index.md`: the callout claiming `docs/AGENTS.md`'s structure block "stops at ADR-3" was itself stale (#862) - true when the page was written (2026-07-18 14:37, commit `c8bb3c9`) but false again ~90 minutes later the same day once `docs/AGENTS.md` was corrected (commit `6d8ead5`). Reworded to a general, still-true trap (the structure block *can* drift behind `docs/ADRs/`, it has before) instead of a point-in-time claim that expires.
 - **Added** - `decisions/scripts-folder-removed.md`: root `scripts/` (106 tracked files, re-verified against `origin/main` rather than the issue's now-stale 98) and root `package.json`/`package-lock.json` deleted per #791 discussion, expanded beyond the issue's 16 orphaned scripts to the whole persisted-script convention. Live-verification scripts are now scratch-only, never committed.
 - **Updated** - `process/deploy-verify-flow.md` and `process/live-playwright-scripts.md`: smoke script section and the shared-helper page rewritten for the scratch-directory approach (no more `scripts/lib/live-browser.mjs`); `process/index.md` description updated to match. `gotchas/sandbox-limitations.md`'s Related line reworded ("browser helper" -> "launch args") to match.
